Update library to use Double for decimal types
This commit depends on an update to the core library to 1.0.0-RC2

+++ src/main/java/org/spdx/library/model/v3_0_1/security/ExternalCvssV2VulnAssessmentRelationship.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2024 Source Auditor Inc.
+ * Copyright (c) 2025 Source Auditor Inc.
  *
  * SPDX-License-Identifier: Apache-2.0
  * 
@@ -228,21 +228,21 @@ public class ExternalCvssV2VulnAssessmentRelationship extends CvssV2VulnAssessme
 	public CvssV2VulnAssessmentRelationship setRelationshipType(@Nullable RelationshipType relationshipType) throws InvalidSPDXAnalysisException {
 		throw new InvalidSPDXAnalysisException("External elements can not set properties");
 	}
-	/**
-	 * @return the score
-	 */
-	public @Nullable Integer getScore() throws InvalidSPDXAnalysisException {
-		return null;
-	}
-	
-	/**
-	 * @param score the score to set
-	 * @return this to chain setters
-	 * @throws InvalidSPDXAnalysisException 
-	 */
-	public CvssV2VulnAssessmentRelationship setScore(@Nullable Integer score) throws InvalidSPDXAnalysisException {
-		throw new InvalidSPDXAnalysisException("External elements can not set properties");
-	}
+    /**
+     * @return the score
+     */
+    public @Nullable Double getScore() throws InvalidSPDXAnalysisException {
+        return null;
+    }
+
+    /**
+     * @param score the score to set
+     * @return this to chain setters
+     * @throws InvalidSPDXAnalysisException
+     */
+    public CvssV2VulnAssessmentRelationship setScore(@Nullable Double score) throws InvalidSPDXAnalysisException {
+        throw new InvalidSPDXAnalysisException("External elements can not set properties");
+    }
 	/**
 	 * @return the vectorString
 	 */
